backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6

# Authentication
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
google-auth>=2.23.0

# Database
pyodbc>=5.0.0

# AI/ML
openai>=1.3.0
whisper>=1.1.10
ollama>=0.1.7
sentence-transformers>=2.2.2

# Performance & Monitoring
redis>=5.0.0
celery>=5.3.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For TestClient

# Development tools
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.18.0